+++ test/libsolidity/smtCheckerTests/functions/constructor_hierarchy_mixed_chain_empty_base.sol
@@ -1,0 +1,24 @@
+pragma experimental SMTChecker;
+contract F {
+	uint a;
+	constructor() public {
+		a = 2;
+	}
+}
+
+contract E is F {}
+contract D is E {
+	constructor() public {
+		a = 3;
+	}
+}
+contract C is D {}
+contract B is C {
+	constructor() public {
+		assert(a == 3);
+		a = 4;
+	}
+}
+
+contract A is B {
+}