+++ test/libsolidity/smtCheckerTests/types/data_location_in_function_type.sol
@@ -1,0 +1,5 @@
+pragma experimental SMTChecker;
+library L {
+    struct Nested { uint y; }
+    function c(function(Nested memory) external returns (uint)[] storage) external pure {}
+}