+++ test/libsolidity/smtCheckerTests/types/mapping_and_array_of_functions.sol
@@ -1,0 +1,7 @@
+pragma experimental SMTChecker;
+contract test {
+    mapping (address => function() internal returns (uint)) a;
+    mapping (address => function() external) b;
+    mapping (address => function() external[]) c;
+    function() external[] d;
+}